REFACTOR: modify header component to handle header icon state logic in hooks
